+++ error-prone-contrib/src/test/resources/tech/picnic/errorprone/bugpatterns/AssertJNumberTemplatesTestOutput.java
@@ -80,10 +80,10 @@ final class AssertJNumberTemplatesTest implements RefasterTemplateTestCase {
     return ImmutableSet.of(
         assertThat((byte) 1).isOdd(),
         assertThat(Byte.valueOf((byte) 1)).isOdd(),
-        assertThat((int) 1).isOdd(),
+        assertThat(1).isOdd(),
         assertThat(Integer.valueOf(1)).isOdd(),
-        assertThat((long) 1).isOdd(),
-        assertThat(Long.valueOf(1L)).isOdd(),
+        assertThat(1L).isOdd(),
+        assertThat(Long.valueOf(1)).isOdd(),
         assertThat((short) 1).isOdd(),
         assertThat(Short.valueOf((short) 1)).isOdd());
   }
@@ -92,10 +92,10 @@ final class AssertJNumberTemplatesTest implements RefasterTemplateTestCase {
     return ImmutableSet.of(
         assertThat((byte) 1).isEven(),
         assertThat(Byte.valueOf((byte) 1)).isEven(),
-        assertThat((int) 1).isEven(),
+        assertThat(1).isEven(),
         assertThat(Integer.valueOf(1)).isEven(),
-        assertThat((long) 1).isEven(),
-        assertThat(Long.valueOf(1L)).isEven(),
+        assertThat(1L).isEven(),
+        assertThat(Long.valueOf(1)).isEven(),
         assertThat((short) 1).isEven(),
         assertThat(Short.valueOf((short) 1)).isEven());
   }
